railhubz_0.05.0-compliepbaisc_path_runz_ish_

+++ sorce_railhubz/base_class_and_functions/enity_1test.h
@@ -48,7 +48,7 @@ double train_y;
     bool arived = false;
 
 public :
-   void load_data(double priority, int spd, Defined_train_path* path_take);
+   void load_data(double priority, float spd, Defined_train_path* path_take);
 
    void update();
